refactor(tests): extract shared session test helpers into sessionTestHelpers.js
Token optimization E: 7 helpers duplicati estratti in shared module.
4 test file aggiornati. ~210 LOC duplicazione rimossa, 150 LOC shared.
254/254 test verdi.
