+++ tests/api/sessionRoundModelEquivalence.test.js
@@ -19,102 +19,16 @@ process.env.IDEA_ENGINE_DISABLE_STATUS_REFRESH = '1';
 
 const test = require('node:test');
 const assert = require('node:assert/strict');
-const request = require('supertest');
 const { createApp } = require('../../apps/backend/app');
-
-// ─────────────────────────────────────────────────────────────────
-// Helpers
-// ─────────────────────────────────────────────────────────────────
-
-function createFlaggedApp(flagValue) {
-  const prior = process.env.USE_ROUND_MODEL;
-  process.env.USE_ROUND_MODEL = flagValue;
-  const handle = createApp({ databasePath: null });
-  return {
-    ...handle,
-    restore: () => {
-      if (prior === undefined) delete process.env.USE_ROUND_MODEL;
-      else process.env.USE_ROUND_MODEL = prior;
-    },
-  };
-}
-
-async function startSession(app, units) {
-  const res = await request(app).post('/api/session/start').send({ units }).expect(200);
-  return res.body.session_id;
-}
-
-async function playerAttack(app, sessionId, actorId, targetId) {
-  return request(app)
-    .post('/api/session/action')
-    .send({ session_id: sessionId, actor_id: actorId, action_type: 'attack', target_id: targetId });
-}
-
-async function turnEnd(app, sessionId) {
-  return request(app).post('/api/session/turn/end').send({ session_id: sessionId });
-}
-
-async function getState(app, sessionId) {
-  const res = await request(app).get('/api/session/state').query({ session_id: sessionId });
-  return res.body;
-}
-
-// Standard 2-unit fixture: player p1 + SIS sis
-function twoUnits(overrides = {}) {
-  return [
-    {
-      id: 'p1',
-      species: 'velox',
-      job: 'skirmisher',
-      hp: overrides.p1Hp != null ? overrides.p1Hp : 10,
-      max_hp: overrides.p1MaxHp != null ? overrides.p1MaxHp : 10,
-      ap: 2,
-      attack_range: 2,
-      initiative: 14,
-      position: overrides.p1Pos || { x: 2, y: 2 },
-      controlled_by: 'player',
-      status: overrides.p1Status || {},
-    },
-    {
-      id: 'sis',
-      species: 'carapax',
-      job: 'vanguard',
-      hp: overrides.sisHp != null ? overrides.sisHp : 10,
-      max_hp: overrides.sisMaxHp != null ? overrides.sisMaxHp : 10,
-      ap: 2,
-      attack_range: overrides.sisRange || 1,
-      initiative: 10,
-      position: overrides.sisPos || { x: 3, y: 2 },
-      controlled_by: 'sistema',
-      status: overrides.sisStatus || {},
-    },
-  ];
-}
-
-// Run a scenario on both flag-off and flag-on, return both results
-async function runDual(units, scenario) {
-  // Flag OFF (legacy)
-  const offApp = createFlaggedApp('false');
-  let offResult;
-  try {
-    offResult = await scenario(offApp.app, units);
-  } finally {
-    offApp.restore();
-    if (typeof offApp.close === 'function') await offApp.close().catch(() => {});
-  }
-
-  // Flag ON (round)
-  const onApp = createFlaggedApp('true');
-  let onResult;
-  try {
-    onResult = await scenario(onApp.app, units);
-  } finally {
-    onApp.restore();
-    if (typeof onApp.close === 'function') await onApp.close().catch(() => {});
-  }
-
-  return { off: offResult, on: onResult };
-}
+const {
+  createFlaggedApp,
+  startSession,
+  playerAttack,
+  turnEnd,
+  getState,
+  twoUnits,
+  runDual,
+} = require('./sessionTestHelpers');
 
 // ─────────────────────────────────────────────────────────────────
 // M6-M8 Batch 1: Init, single attack, single move
