+++ tests/api/sessionTurnEndWrapper.test.js
@@ -12,48 +12,7 @@ const test = require('node:test');
 const assert = require('node:assert/strict');
 const request = require('supertest');
 const { createApp } = require('../../apps/backend/app');
-
-function withRoundFlag(value) {
-  const prior = process.env.USE_ROUND_MODEL;
-  process.env.USE_ROUND_MODEL = value;
-  return () => {
-    if (prior === undefined) delete process.env.USE_ROUND_MODEL;
-    else process.env.USE_ROUND_MODEL = prior;
-  };
-}
-
-async function startSession(app, units) {
-  const res = await request(app)
-    .post('/api/session/start')
-    .send({
-      units: units || [
-        {
-          id: 'p1',
-          species: 'velox',
-          job: 'skirmisher',
-          hp: 10,
-          ap: 2,
-          attack_range: 2,
-          initiative: 14,
-          position: { x: 2, y: 2 },
-          controlled_by: 'player',
-        },
-        {
-          id: 'sis',
-          species: 'carapax',
-          job: 'vanguard',
-          hp: 10,
-          ap: 2,
-          attack_range: 1,
-          initiative: 10,
-          position: { x: 3, y: 2 },
-          controlled_by: 'sistema',
-        },
-      ],
-    })
-    .expect(200);
-  return res.body.session_id;
-}
+const { withRoundFlag, startSession } = require('./sessionTestHelpers');
 
 // ─────────────────────────────────────────────────────────────────
 // Flag on — /turn/end via round flow
